requirements.txt
annotated-types==0.7.0
anyio==4.9.0
app==0.0.1
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distlib==0.3.9
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
greenlet==3.2.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jose==1.0.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
mysql-connector-python==9.3.0
numpy==2.2.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
platformdirs==4.3.7
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
pytweening==1.2.0
pytz==2025.2
rarfile==4.2
redis==5.2.1
requests==2.32.3
selenium==4.31.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
spotipy==2.25.1
SQLAlchemy==2.0.40
starlette==0.46.2
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.2
virtualenv==20.30.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yt-dlp==2025.3.27
